+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/lightspeed/trex/stage_manager/plugin/interaction/usd/all_meshes.py
@@ -40,6 +40,7 @@ class AllMeshesInteractionPlugin(_StageManagerUSDInteractionPlugin):
             "IgnorePrimsFilterPlugin",
             "IsCaptureFilterPlugin",
             "IsCategoryFilterPlugin",
+            "VisiblePrimsFilterPlugin",
             "LightPrimsFilterPlugin",
             "MeshPrimsFilterPlugin",
             "OmniPrimsFilterPlugin",