+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_prims.py
@@ -38,6 +38,7 @@ class AllPrimsInteractionPlugin(_StageManagerUSDInteractionPlugin):
             "IsCaptureFilterPlugin",
             "IsCategoryFilterPlugin",
             "RemixLogicPrimsFilterPlugin",
+            "VisiblePrimsFilterPlugin",
             "LightPrimsFilterPlugin",
             "MaterialPrimsFilterPlugin",
             "MeshPrimsFilterPlugin",