+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.4.0]
+### Added
+- Added `VisiblePrimsFilterPlugin` to the interactions plugins
+
 ## [2.3.0]
 ### Added
 - Added `DeleteRestoreActionWidgetPlugin`